Add `NoiseModel.from_backend` static method
* The functionality of basic_device_noise_model has been moved to `NoiseModel.from_backend`
* Using `basic_device_noise_model` is deprecated.
* Improved NoiseModel API Docs
