+++ qiskit/providers/aer/noise/__init__.py
@@ -10,7 +10,7 @@
 # copyright notice, and modified files need to carry a notice indicating
 # that they have been altered from the originals.
 
-"""
+r"""
 ================================================
 Noise Models (:mod:`qiskit.providers.aer.noise`)
 ================================================
@@ -19,6 +19,69 @@ Noise Models (:mod:`qiskit.providers.aer.noise`)
 
 This module contains classes and functions to build a noise model for
 simulating a Qiskit quantum circuit in the presence of errors.
+
+
+Building Noise Models
+=====================
+
+The :class:`NoiseModel` class is used to represent noise model for the
+:class:`~qiskit.providers.aer.QasmSimulator`. It can be used to construct
+custom noise models for simulator, to to automatically generate a basic
+device noise model for an IBMQ backend.
+
+
+Device Noise Models
+-------------------
+
+A simplified approximate :class:`NoiseModel` can be generated automatically
+from the properties of real device backends from the IBMQ provider using the
+:meth:`~NoiseModel.from_backend` method. See the method documentation for
+details.
+
+**Example**
+
+Constructing an approximate noise model from a backend
+
+.. code-block:: python
+
+    from qiskit import IBMQ
+    from qiskit.providers.aer.noise import NoiseModel
+
+    provider = IBMQ.load_account()
+    backend = provider.get_backend('ibmq_vigo')
+    noise_model = NoiseModel.from_backend(backend)
+
+
+Custom Noise Models
+-------------------
+
+Custom noise models can be used by adding :class:`QuantumError` to circuit
+gate, reset or measure instructions, and :class:`ReadoutError` to measure
+instructions. This module includes several helper functions for generating
+:class:`QuantumError` instances based on canonical error models used in
+Quantum Information Theory that can simplify building noise models. See the
+documentation for the :class:`NoiseModel` class for additional details.
+
+**Example**
+
+Constructing a simple depolarizing noise model
+
+.. code-block:: python
+
+    import qiskit.providers.aer.noise as noise
+
+    # Error probabilities
+    prob_1 = 0.001  # 1-qubit gate
+    prob_2 = 0.01   # 2-qubit gate
+
+    # Depolarizing quantum errors
+    error_1 = noise.depolarizing_error(prob_1, 1)
+    error_2 = noise.depolarizing_error(prob_2, 2)
+
+    # Add errors to noise model
+    noise_model = noise.NoiseModel
+    noise_model.add_all_qubit_quantum_error(error_1, ['u1', 'u2', 'u3'])
+    noise_model.add_all_qubit_quantum_error(error_2, [cx'])
 
 
 Classes
@@ -56,18 +119,22 @@ The following functions can be used to generate many common types of
     kraus_error
 
 
-Noise Model Functions
-=====================
+Device Noise Parameters
+=======================
 
-The following functions can be used to generate approximate noise models for
-IBMQ hardware devices based on the parameters in their backend properties.
+The following are utility functions which can be used for extracting error
+parameters and error objects from device `BackendProperties`.
 
 .. autosummary::
     :toctree: ../stubs/
 
-    basic_device_noise_model
-    basic_device_readout_errors
-    basic_device_gate_errors
+    device.basic_device_readout_errors
+    device.basic_device_gate_errors
+    device.gate_param_values
+    device.gate_error_values
+    device.gate_length_values
+    device.readout_error_values
+    device.thermal_relaxation_values
 """
 
 # Noise and Error classes
@@ -86,11 +153,6 @@ from .errors import thermal_relaxation_error
 from .errors import phase_amplitude_damping_error
 from .errors import amplitude_damping_error
 from .errors import phase_damping_error
-
-# Noise model generating functions
-from .device.models import basic_device_noise_model
-from .device.models import basic_device_readout_errors
-from .device.models import basic_device_gate_errors
 
 # Submodules
 from . import errors
