+++ qiskit/providers/aer/noise/noise_model.py
@@ -17,12 +17,17 @@ import json
 import logging
 
 from qiskit.circuit import Instruction
+from qiskit.providers import BaseBackend
+from qiskit.providers.models import BackendProperties
+
 from qiskit.providers.aer.backends.aerbackend import AerJSONEncoder
 from qiskit.providers.aer.backends.qasm_simulator import QasmSimulator
 
 from .noiseerror import NoiseError
 from .errors.quantum_error import QuantumError
 from .errors.readout_error import ReadoutError
+from .device.models import basic_device_gate_errors
+from .device.models import basic_device_readout_errors
 
 logger = logging.getLogger(__name__)
 
@@ -115,6 +120,126 @@ class NoiseModel:
     def noise_qubits(self):
         """Return the set of noisy qubits for this noise model."""
         return sorted(self._noise_qubits)
+
+    @classmethod
+    def from_backend(cls, backend,
+                     gate_error=True,
+                     readout_error=True,
+                     thermal_relaxation=True,
+                     temperature=0,
+                     gate_lengths=None,
+                     standard_gates=True):
+        """Return a noise model derived from a devices backend properties.
+
+        This function generates a noise model based on:
+
+        * 1 and 2 qubit gate errors consisting of a
+          :func:`depolarizing_error` followed
+          by a :func:`thermal_relaxation_error`.
+
+        * Single qubit :class:`ReadoutError` on all measurements.
+
+        The Error error parameters are tuned for each individual qubit based on
+        the :math:`T_1`, :math:`T_2`, frequency and readout error parameters for
+        each qubit, and the gate error and gate time parameters for each gate
+        obtained from the device backend properties.
+
+        **Additional Information**
+
+        The noise model includes the following errors:
+
+        * If ``readout_error=True`` include single qubit readout
+          errors on measurements.
+
+        * If ``gate_error=True`` and ``thermal_relaxation=True`` include:
+
+            * Single-qubit gate errors consisting of a :func:`depolarizing_error`
+              followed by a :func:`thermal_relaxation_error` for the qubit the
+              gate acts on.
+
+            * Two-qubit gate errors consisting of a 2-qubit
+              :func:`depolarizing_error` followed by single qubit
+              :func:`thermal_relaxation_error` on each qubit participating in
+              the gate.
+
+        * If ``gate_error=True`` is ``True`` and ``thermal_relaxation=False``:
+
+            * An N-qubit :func:`depolarizing_error` on each N-qubit gate.
+
+        * If ``gate_error=False`` and ``thermal_relaxation=True`` include
+          single-qubit :func:`thermal_relaxation_errors` on each qubits
+          participating in a multi-qubit gate.
+
+        For best practice in simulating a backend make sure that the
+        circuit is compiled using the set of basis gates in the noise
+        module by setting ``basis_gates=noise_model.basis_gates``
+        and using the device coupling map with
+        ``coupling_map=backend.configuration().coupling_map``
+
+        **Specifying custom gate times**
+
+        The ``gate_lengths`` kwarg can be used to specify custom gate times
+        to add gate errors using the :math:`T_1` and :math:`T_2` values from
+        the backend properties. This should be passed as a list of tuples
+        ``gate_lengths=[(name, value), ...]``
+        where ``name`` is the gate name string, and ``value`` is the gate time
+        in nanoseconds.
+
+        If a custom gate is specified that already exists in
+        the backend properties, the ``gate_lengths`` value will override the
+        gate time value from the backend properties.
+        If non-default values are used gate_lengths should be a list
+
+        Args:
+            backend (Backend or BackendProperties): backend properties.
+            gate_error (bool): Include depolarizing gate errors (Default: True).
+            readout_error (Bool): Include readout errors in model
+                                  (Default: True).
+            thermal_relaxation (Bool): Include thermal relaxation errors
+                                       (Default: True).
+            temperature (double): qubit temperature in milli-Kelvin (mK) for
+                                  thermal relaxation errors (Default: 0).
+            gate_lengths (list): Custom gate times for thermal relaxation errors.
+                                  Used to extend or override the gate times in
+                                  the backend properties (Default: None))
+            standard_gates (bool): If true return errors as standard
+                                   qobj gates. If false return as unitary
+                                   qobj instructions (Default: True)
+
+        Returns:
+            NoiseModel: An approximate noise model for the device backend.
+
+        Raises:
+            NoiseError: If the input backend is not valid.
+        """
+        if isinstance(backend, BaseBackend):
+            properties = backend.properties()
+            if not properties:
+                raise NoiseError('Qiskit backend {} does not have a '
+                                 'BackendProperties'.format(backend))
+        elif isinstance(backend, BackendProperties):
+            properties = backend
+        else:
+            raise NoiseError('{} is not a Qiskit backend or'
+                             ' BackendProperties'.format(backend))
+        noise_model = NoiseModel()
+
+        # Add single-qubit readout errors
+        if readout_error:
+            for qubits, error in basic_device_readout_errors(properties):
+                noise_model.add_readout_error(error, qubits)
+
+        # Add gate errors
+        gate_errors = basic_device_gate_errors(
+            properties,
+            gate_error=gate_error,
+            thermal_relaxation=thermal_relaxation,
+            gate_lengths=gate_lengths,
+            temperature=temperature,
+            standard_gates=standard_gates)
+        for name, qubits, error in gate_errors:
+            noise_model.add_quantum_error(error, name, qubits)
+        return noise_model
 
     def __repr__(self):
         """Display noise model"""
